[#23] Refactor: QueryFeedBySortOption 삭제 후 피드별 정렬 조회 로직을 FeedRepository로 이동
